Update -Elastic Search Fixed Name property "Id" to "ID" -EfRepository fixed update entity when entity state is not equal detached -Remove unusing property in ca view model -Change parameter name in constant value

+++ DS.Bll/AttachmentBll.cs
@@ -91,7 +91,7 @@ namespace DS.Bll
             {
                 result = true;
                 var attach = _unitOfWork.GetRepository<DS.Data.Pocos.Attachment>().GetById(item.ID);
-                if (File.Exists(Path.Combine(documentPath,attach.SavedFileName)))
+                if (attach != null && File.Exists(Path.Combine(documentPath,attach.SavedFileName)))
                 {
                     //Remove file in local storage.
                     File.Delete(Path.Combine(documentPath, attach.SavedFileName));
